chg: commands in separate executable files
